% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcGmm.R
\name{tcGmm}
\alias{tcGmm}
\title{Fit Two-Condition Gaussian Mixture Model}
\usage{
tcGmm(y, g, zInit, maxIter = 100, thresh = 1e-08, verboseN = TRUE)
}
\arguments{
\item{y}{A matrix with rows as samples (cells) and columns as features (genes)}

\item{g}{A vector indicating condition 1 (0) and condition 2 (1)}

\item{zInit}{A matrix indicating the assignment of groups with rows as samples and columns as groups}

\item{maxIter}{A numeric value of maximum iteration number. Default is 100.}

\item{thresh}{A numeric value of the converge criteria. Define as the Frobenius norm of the difference of current mean and mean in last iteration. Default is 1e-8.}

\item{verboseN}{A llogical value. Whether to print the iteration number.}
}
\value{
A list which contains the fitting result.
## mu The mean parameter
## sigma The standard deviation parameter
## delta The shift of mean parameter
## z The assignment of groups
## model The fitted regression model of each group
}
\description{
This function fits the Two-Condition Gaussian Mixture Model (TCGMM).
We assume that the latent groups are consistent between two conditions only with shifts in mean parameters.
}
\author{
Dongyuan Song
}
